Cleaned up states, made statistics not static, organized all states and transitions
