+++ ihmc-high-level-behaviors/src/main/java/us/ihmc/behaviors/activeMapping/ContinuousHikingStateMachine/DoNothingState.java
@@ -1,10 +1,7 @@
-package us.ihmc.behaviors.activeMapping.ContinuousHikingStatesAndTransitions;
+package us.ihmc.behaviors.activeMapping.ContinuousHikingStateMachine;
 
-import behavior_msgs.msg.dds.ContinuousWalkingCommandMessage;
 import controller_msgs.msg.dds.PauseWalkingMessage;
-import us.ihmc.behaviors.activeMapping.ContinuousHikingParameters;
 import us.ihmc.behaviors.activeMapping.ContinuousPlanner;
-import us.ihmc.behaviors.activeMapping.ControllerFootstepQueueMonitor;
 import us.ihmc.communication.HumanoidControllerAPI;
 import us.ihmc.communication.ros2.ROS2Helper;
 import us.ihmc.euclid.referenceFrame.FramePose3D;
@@ -12,14 +9,9 @@ import us.ihmc.euclid.referenceFrame.ReferenceFrame;
 import us.ihmc.footstepPlanning.monteCarloPlanning.TerrainPlanningDebugger;
 import us.ihmc.humanoidRobotics.frames.HumanoidReferenceFrames;
 import us.ihmc.log.LogTools;
-import us.ihmc.perception.heightMap.TerrainMapData;
 import us.ihmc.robotics.robotSide.RobotSide;
 import us.ihmc.robotics.stateMachine.core.State;
 import us.ihmc.ros2.ROS2PublisherBasics;
-
-import java.util.concurrent.atomic.AtomicReference;
-
-import static us.ihmc.behaviors.activeMapping.ContinuousPlannerSchedulingTask.statistics;
 
 public class DoNothingState implements State
 {
@@ -42,11 +34,10 @@ public class DoNothingState implements State
       pauseWalkingPublisher = ros2Helper.getROS2NodeInterface().createPublisher(HumanoidControllerAPI.getTopic(PauseWalkingMessage.class, simpleRobotName));
    }
 
-
    @Override
    public void onEntry()
    {
-      LogTools.warn("Entering [DO_NOTHING] state");
+      LogTools.warn(String.format("Entering %s", getClass().getSimpleName()));
    }
 
    @Override
@@ -54,6 +45,7 @@ public class DoNothingState implements State
    {
       PauseWalkingMessage message = new PauseWalkingMessage();
 
+      // This state can be entered when we want to stop Continuous Hiking, if that's the cause, pause walking so the robot stops
       if (continuousPlanner.isInitialized())
       {
          message.setPause(true);
